Add class SmartRunner, which implements the default "smart" Builder behavior.
With this, the old Builder.runner() class member is superfluous as its
functionality is provided by the callable instance attribute runner().
Instead, the runner behavior should be set by changing the Builder's
runner instance attribute.

However, for backwards compatibility, we have to allow for the
possibility that a runner() method was provided by a derived class.
Therefore, do not set the "runner" attribute if one is already
defined.

git-svn-id: http://fabricate.googlecode.com/svn/trunk@60 7d55530a-7836-11de-8a01-315fa775be5f