Updated TrackerRunner to correctly convert the file operations to relevant file names for the dependencies
Works with current tests in git bash except rename and symlink.

Rename does not track the input file or the output file unless the output file already exists. If the output already exists then it is incorrectly marked as in input file.

Symlink tracking does not work because symlinks are not really supported in git bash or windows.